Rename the functions base on clean code
